fix issue with deleting hover images after edit

+++ server/controllers/admin.js
@@ -121,7 +121,7 @@ exports.updateProduct = (req, res, next) => {
         if (prod.images.hover) fileHelper.deleteFile(prod.images.hover);
         const hoverImage = req.files['hover'][0];
         prod.images.hover = hoverImage ? hoverImage.filename : null;
-      }else if(prod.images.hover){
+      } else if (prod.images.hover && req.body.removeHoverImage) {
         fileHelper.deleteFile(prod.images.hover);
         prod.images.hover = null;
       }
